# Conflicts:
#	integrationTests/singleShard/block/executingMiniblocksSc_test.go

+++ consensus/round/round.go
@@ -9,7 +9,7 @@ import (
 
 // round defines the data needed by the rounder
 type round struct {
-	index        int32         // represents the index of the round in the current chronology (current time - genesis time) / round duration
+	index        int64         // represents the index of the round in the current chronology (current time - genesis time) / round duration
 	timeStamp    time.Time     // represents the start time of the round in the current chronology genesis time + round index * round duration
 	timeDuration time.Duration // represents the duration of the round in current chronology
 	syncTimer    ntp.SyncTimer
@@ -36,7 +36,7 @@ func NewRound(
 func (rnd *round) UpdateRound(genesisTimeStamp time.Time, currentTimeStamp time.Time) {
 	delta := currentTimeStamp.Sub(genesisTimeStamp).Nanoseconds()
 
-	index := int32(math.Floor(float64(delta) / float64(rnd.timeDuration.Nanoseconds())))
+	index := int64(math.Floor(float64(delta) / float64(rnd.timeDuration.Nanoseconds())))
 
 	if rnd.index != index {
 		rnd.index = index
@@ -45,7 +45,7 @@ func (rnd *round) UpdateRound(genesisTimeStamp time.Time, currentTimeStamp time.
 }
 
 // Index returns the index of the round in current epoch
-func (rnd *round) Index() int32 {
+func (rnd *round) Index() int64 {
 	return rnd.index
 }
