+++ process/interface.go
@@ -17,7 +17,7 @@ import (
 
 // TransactionProcessor is the main interface for transaction execution engine
 type TransactionProcessor interface {
-	ProcessTransaction(transaction *transaction.Transaction, round uint32) error
+	ProcessTransaction(transaction *transaction.Transaction, round uint64) error
 }
 
 // SmartContractResultProcessor is the main interface for smart contract result execution engine
@@ -40,11 +40,11 @@ type TransactionCoordinator interface {
 	RestoreBlockDataFromStorage(body block.Body) (int, map[int][][]byte, error)
 	RemoveBlockDataFromPool(body block.Body) error
 
-	ProcessBlockTransaction(body block.Body, round uint32, haveTime func() time.Duration) error
+	ProcessBlockTransaction(body block.Body, round uint64, haveTime func() time.Duration) error
 
 	CreateBlockStarted()
-	CreateMbsAndProcessCrossShardTransactionsDstMe(header data.HeaderHandler, maxTxSpaceRemained uint32, maxMbSpaceRemained uint32, round uint32, haveTime func() bool) (block.MiniBlockSlice, uint32, bool)
-	CreateMbsAndProcessTransactionsFromMe(maxTxSpaceRemained uint32, maxMbSpaceRemained uint32, round uint32, haveTime func() bool) block.MiniBlockSlice
+	CreateMbsAndProcessCrossShardTransactionsDstMe(header data.HeaderHandler, maxTxSpaceRemained uint32, maxMbSpaceRemained uint32, round uint64, haveTime func() bool) (block.MiniBlockSlice, uint32, bool)
+	CreateMbsAndProcessTransactionsFromMe(maxTxSpaceRemained uint32, maxMbSpaceRemained uint32, round uint64, haveTime func() bool) block.MiniBlockSlice
 
 	CreateMarshalizedData(body block.Body) (map[uint32]block.MiniBlockSlice, map[string][][]byte)
 
@@ -56,8 +56,8 @@ type TransactionCoordinator interface {
 // SmartContractProcessor is the main interface for the smart contract caller engine
 type SmartContractProcessor interface {
 	ComputeTransactionType(tx *transaction.Transaction) (TransactionType, error)
-	ExecuteSmartContractTransaction(tx *transaction.Transaction, acntSrc, acntDst state.AccountHandler, round uint32) error
-	DeploySmartContract(tx *transaction.Transaction, acntSrc state.AccountHandler, round uint32) error
+	ExecuteSmartContractTransaction(tx *transaction.Transaction, acntSrc, acntDst state.AccountHandler, round uint64) error
+	DeploySmartContract(tx *transaction.Transaction, acntSrc state.AccountHandler, round uint64) error
 }
 
 // IntermediateTransactionHandler handles transactions which are not resolved in only one step
@@ -79,14 +79,14 @@ type PreProcessor interface {
 	RestoreTxBlockIntoPools(body block.Body, miniBlockPool storage.Cacher) (int, map[int][]byte, error)
 	SaveTxBlockToStorage(body block.Body) error
 
-	ProcessBlockTransactions(body block.Body, round uint32, haveTime func() time.Duration) error
+	ProcessBlockTransactions(body block.Body, round uint64, haveTime func() time.Duration) error
 	RequestBlockTransactions(body block.Body) int
 
 	CreateMarshalizedData(txHashes [][]byte) ([][]byte, error)
 
 	RequestTransactionsForMiniBlock(mb block.MiniBlock) int
-	ProcessMiniBlock(miniBlock *block.MiniBlock, haveTime func() bool, round uint32) error
-	CreateAndProcessMiniBlock(sndShardId, dstShardId uint32, spaceRemained int, haveTime func() bool, round uint32) (*block.MiniBlock, error)
+	ProcessMiniBlock(miniBlock *block.MiniBlock, haveTime func() bool, round uint64) error
+	CreateAndProcessMiniBlock(sndShardId, dstShardId uint32, spaceRemained int, haveTime func() bool, round uint64) (*block.MiniBlock, error)
 
 	GetAllCurrentUsedTxs() map[string]data.TransactionHandler
 }
@@ -96,9 +96,9 @@ type BlockProcessor interface {
 	ProcessBlock(blockChain data.ChainHandler, header data.HeaderHandler, body data.BodyHandler, haveTime func() time.Duration) error
 	CommitBlock(blockChain data.ChainHandler, header data.HeaderHandler, body data.BodyHandler) error
 	RevertAccountState()
-	CreateBlockBody(round uint32, haveTime func() bool) (data.BodyHandler, error)
+	CreateBlockBody(round uint64, haveTime func() bool) (data.BodyHandler, error)
 	RestoreBlockIntoPools(header data.HeaderHandler, body data.BodyHandler) error
-	CreateBlockHeader(body data.BodyHandler, round uint32, haveTime func() bool) (data.HeaderHandler, error)
+	CreateBlockHeader(body data.BodyHandler, round uint64, haveTime func() bool) (data.HeaderHandler, error)
 	MarshalizedDataToBroadcast(header data.HeaderHandler, body data.BodyHandler) (map[uint32][]byte, map[string][][]byte, error)
 	DecodeBlockBody(dta []byte) data.BodyHandler
 	DecodeBlockHeader(dta []byte) data.HeaderHandler
@@ -233,7 +233,7 @@ type TopicMessageHandler interface {
 // ChronologyValidator defines the functionality needed to validate a received header block (shard or metachain)
 // from chronology point of view
 type ChronologyValidator interface {
-	ValidateReceivedBlock(shardID uint32, epoch uint32, nonce uint64, round uint32) error
+	ValidateReceivedBlock(shardID uint32, epoch uint32, nonce uint64, round uint64) error
 }
 
 // DataPacker can split a large slice of byte slices in smaller packets
@@ -246,8 +246,8 @@ type BlocksTracker interface {
 	UnnotarisedBlocks() []data.HeaderHandler
 	RemoveNotarisedBlocks(headerHandler data.HeaderHandler) error
 	AddBlock(headerHandler data.HeaderHandler)
-	SetBlockBroadcastRound(nonce uint64, round int32)
-	BlockBroadcastRound(nonce uint64) int32
+	SetBlockBroadcastRound(nonce uint64, round int64)
+	BlockBroadcastRound(nonce uint64) int64
 }
 
 // RequestHandler defines the methods through which request to data can be made
